9/21. Finished the decimal to binary conversation part
Imports math library, which will be used to convert the decimal number to binary.  the user input of what number they want to convert. Converts the input, which is a string, to an int. This will allow it to be used later for a math function. Converts the decimal number to binary and saves it in the variable binaryVersion. The bin function adds a random "b" in the binary number. The .replace() removes the b and saves the new value in binaryVersion1